requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosmtplib==2.0.2
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
dataclasses-json==0.6.7
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.104.1
fastapi-mail==1.4.1
frozenlist==1.6.0
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.0
google-auth-httplib2==0.2.0
google-generativeai==0.3.1
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.9
langchain-community==0.0.38
langchain-core==0.1.53
langchain-google-genai==0.0.9
langsmith==0.1.147
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.4.3
mypy_extensions==1.1.0
numpy==1.26.4
orjson==3.10.18
packaging==23.2
passlib==1.7.4
pdf2image==1.17.0
pillow==10.2.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.7
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
pymongo==4.6.1
pyparsing==3.2.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
tenacity==8.5.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.24.0
yarl==1.20.0
